pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
pyright = "^1.1.403"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
